refactor fixes and organizing components
