+++ src/routes/signup/1_Introduction.svelte
@@ -1,0 +1,27 @@
+<h5 class="text-center h4 my-7">This app is a little different from what you're used to.</h5>
+<div class="grid place-content-center text-justify text-lg mb-10">
+	<p class="text-xl mb-3 italic underline underline-offset-8">
+		So please actually read the following:
+	</p>
+	<p>
+		<b class="text-secondary-500">Discreetly</b> is an <b>anonymous</b> chat app
+	</p>
+	<p>Yes, you are <span class="text-success-500"><i>actually</i> anonymous</span></p>
+	<p>
+		But you can still get <b class="text-secondary-500">banned</b> if you
+		<b class="text-secondary-500">spam</b>, or if you get
+		<b class="text-secondary-500">voted out</b>.
+	</p>
+	<p>
+		There is
+		<b class="text-secondary-500">no unban</b>
+	</p>
+	<p>
+		<b>Backup your identity</b>, there is
+		<i class="text-secondary-500">no account recovery</i> (yet)
+	</p>
+	<p class="mt-2 mb-4">
+		If you want to understand how this works:
+		<a href="/about" class="btn btn-sm variant-soft-tertiary ms-2 mt-2">Read More Here</a>
+	</p>
+</div>
